[#48416887] unicorn check_client_connection
Enable unicorn's `check_client_connection` option for versions of unicorn
that support it (>= 4.5.0). Raise a warning on STDERR for those that don't.
This will allow us to progressively upgrade all projects to make use of
this.

In short, the option will attempt to verify whether a request queued on the
master process still has a client listening before passing it to a worker.
This should reduce the load and recovery time when unicorn becomes swamped.
You can read more about it here:

- http://unicorn.bogomips.org/Unicorn/Configurator.html#method-i-check_client_connection
- http://www.shopify.co.uk/technology/7535298-what-does-your-webserver-do-when-a-user-hits-refresh

The effectiveness of this may be limited whilst we are using HTTPS for
internally reverse-proxed connections because isn't able to detect the
client disconnect so quickly:

- http://nginx.2469901.n2.nabble.com/Dropped-https-client-connection-doesn-t-drop-backend-proxy-pass-connection-td7584346.html

However this is still a step in the right direction and shouldn't cause any
negative effects.