[#47368223] Enable auth on licensify preview
This is a rehash of a previous attempt to do this.

The complication when enabling last time, was that while we want to
protect /apply-for-a-license we do not want to protect /api as it breaks
the application completely.

We are using govuk:app to protect the vhost, but introducing additional
config in an existing snippet file to unprotect /api