[Fixes #49052711] Add openjdk6::jdk to s_support
To satisfy dependencies in `java::set_defaults` below. There isn't currently
a JDK installed on these nodes. But installing one isn't too harmful, and
saves defaulting to Sun6.

This issue doesn't affect nodes using `java::oracle7` because the JRE and
JDK classes are one-and-the-same.

We might want to consider making the `jre`/`jdk` params optional in the
future if we have nodes where we really don't want one or the other
installed.