Adjust how we alert on disk I/O
collectd prior to 5.1.1 misreported disk_time (time in milliseconds spent by the
kernel reading or writing data to disk) by a factor of the collection interval
(in our case, 10s):

  https://github.com/collectd/collectd/issues/150

This adjusts our disk time alerting to take account of this change. In
particular, rather than alerting on the mean value of this metric over the last
30 seconds (which is the default operation of check_graphite), we now check the
median value as measured over the last 5 minutes. This means that we'll miss
short spikes in disk usage (a good thing) and instead only alert on long
persistent high disk usage. The last major event in production which was I/O
related was a runaway virus scanner on the asset-master machine, and the warn
and crit values chosen here are based off the values of this metric from that
incident:

  http://tinyurl.com/cjykle8