Define the destination directory in the es_dump script
By defining the directory in the script, instead of through a
parameter, we can set the expectation that all index dumps will
only be saved in the /var/es_dump directory. This should make it
easier to manage if there comes a time where archives need to
be deleted from boxes.