Remove transformNull() from MySQL repl check
Remove `transformNull()` which was intended to catch scenarios when MySQL
replication had stopped altogether and MySQL reports NULL. The `--droplast`
was used to support this.

It turns out that differentiating between when replication has stopped
(MySQL reports a NULL value) and absent data (which is also NULL, when
collectd hasn't sent or Graphite hasn't stored any data) is impossible. As a
result this was generating a lot of false positives.

We've split the "replication has stopped altogether" scenario out into a
separate check. This will no report is replication is running, but is behind
by N amount. When replication is broken it will flip to `UNKNOWN` and the
`mysql_check_slave` alert will fire as `CRITICAL`.