www_crt/www_key for Vagrant multi-VM
Repurpose the duplicate `wildcard_alphagov_crt` entry to use the same
certificate (and key, new entry) for `certtype => 'www'`. This will fix
errors when bringing up a `cache-X.router` node in Vagrant multi-VM because
an empty value meant that Nginx refused to start.

This may not be the correct certificate to use. But it will do, since we
already have access to it.

/cc @annashipman