[#49910675] Mirrorer handling of 429 responses
We're moving response codes for rate limiting from 503 to 429. The mirrorer
attempts to handle these kind of errors gracefully, by backing off for one
second and then retrying. Maintain this behaviour for the new 429 response.

Test for a handful of the specified codes. Broke handled codes out to a
class constant because munging a range and an integer became a longer line
of code.