Reinstate <1mo retention for Licensify backups
After finding ticket [#36538039] and a discussion with @philpotter, we
decided that this was removed in error. We hadn't realised that it was
called by Licensify in `manifests/classes/elms.pp` with the specific purpose
of limiting backup retention/aggregation of sensitive data.

As a bonus, we get to reclaim some storage.

This reverts commits:
  - a326eb7b4aed3265bd47183d162f0c8004e51542 (`domonthly` param)
  - 4dd7dedda8d89adbd34badc3e9464c0469af56a4 (template)
  - 4089c27f0dcbc6d1292c2f98445550947f6e5639 (licensify caller)

The unused `members` params was removed from the last commit. There are no
references to this in the template. Likewise leaving commit:
  - 1a28f92524c4ae15b47dc699857640101d65b17c