Regex route to performance api for varnish
The performance platform exposes an API for reading and querying the
data stored in our backend. This API is exposed with a URL that matches
the pattern: /performance/{service}/api. So far licensing has been the
only active service, but now we are starting to add more, so we need a
different routing from varnish to the application. We prefer having a
generic regex rather than a whitelist of the active services, because
the services are already whitelisted in the configuration of nginx for
the publicapi vhost. This way we don't have to do the change in two
places every time we add a service.