[#50181413] Match uppercase org acronym FURLs
Match the uppercase counterparts of FURLs that are organisation acronyms. By
splitting them out to a separate hash and using `~*` on the location. This
is in response to some feedback that we've seen for people requesting (e.g.)
`gov.uk/MOD` and getting a 404.

The behaviour is non-deterministic when passed through Akamai. Because we
have enabled case-insensitive cache keys, occasionally the 404 would be
treated as the authoratative result and also be served for lowercase
requests.

This should address that particular inconsistency, at least until we
investigate whether we can disable that Akamai config option (for `www.gov`,
not redirector). Regardless of that, I think it's reasonable that the
acronyms should be available in their uppercase form.