[#48998005] Don't restart Nagios on Nginx changes
Move all Nginx resources out of `::config` and into a new `::nginx`
sub-class so that they don't trigger unwanted restarts of `::service`.

The `::end` anchor hangs off this because we want a calling class to know
that the service may not be completely operational if Nginx fails to
configure.

I have double checked that the moved variables are only referenced by the
one template that has been moved.