[#48998005] Don't restart Nginx on log changes
Move all log related resources out of `::config` and into the existing
`::logging` sub-class so that they don't trigger unwanted restarts of
`::service`.

The `::end` anchor hangs off this because we want a calling class to know
that the service may not be completely operational if Nginx fails to
configure. Also added `::service` for the same reason.