[#48998005] Don't restart Puppet on Nginx changes
Move all Nginx and log related resources out of `::config` and into a new
`::nginx` sub-class so that they don't trigger unwanted restarts of
`::service`.

The `::end` anchor hangs off this because we want a calling class to know
that the service may not be completely operational if Nginx fails to
configure.

I've also flipped the relationship of `::end`/`::service` from `notify` to
`subscribe` because I think it reads a better, inline with the other
`require` params.