[#42296389] Upgrading rbenv module version
This is to address a bug, where bundler is installing to system ruby
because of a previous bug. The previous bug was where shims folder
was not populated as rbenv rehash has to run with RBENV_ROOT. Hence
ruby version was not being correctly picked leading to bundler pick the
default ruby.