Replace custom Nagios plugin with builtin
Don't maintain our own code when we could be using somebody else's. Even if
it is Perl. Perlarrrrgh.

I haven't baked the arguments into `Nrpe_config[]` because we may want to
use different permutations of warning and critical. I also think it's
clearer to define these in the same place as the `Nagios_check[]`.

I'm pretty sure that the argument parsing should work like this. But we
won't know for sure until it hits preview.