Move LB node lists to extdata
Move the list of load balanced nodes out to extdata. This will allow them to
be changed independently of normal Puppet code promotions, and allow us to
specify a different number of nodes in Preview (where we don't need to over
provision). Required by the addition of a new `whitehall-frontend` node in
Staging and Production.

I'm avoiding the use of PuppetDB for now since it would introduce some
quirks. Notably we wouldn't be able to use these nodes in Vagrant without a
Puppetmaster, and it would require a double Puppet run for nodes to come
into service.

Existing nodes committed in gds/deployment@74c9bdc