govuk_spinup: initialize rbenv on spinup
rbenv is installed and working on boxes that need it, but is currently not
active for non-login shells (as in, for example, the processes started by
upstart).

This change will activate rbenv for all applications. Any applications that have
a .ruby-version file will now use that ruby version if available.

NB: If a .ruby-version file specifies a version of ruby that isn't installed on
the production servers, the application will fail to start.