[#50518349] Nagios check_reboot_required
Add a new Nagios check that will alert when a package requires a machine to
be rebooted. This addresses, in particular, cases where a newer kernel
package has been installed by unattended-upgrades but we haven't noticed to
reboot the machine.

As such, I expect this to be a bit noisy when we first roll it out. But it
will give us a checklist of nodes to perform rolling reboots in order to
bring up-to-date with security fixes.