Fix Nagios check_file_age arguments
As predicted in aed99dd, the argument parsing didn't quite work:

    NRPE: Command '691200' not defined
    NRPE: Command '172800' not defined

Enclose the one argument to `check_file_age` with quotes to ensure that it
gets passed over as-is. I've tested this with `check_nrpe` on the CLI.

Also fix the values of `-w` and `-c`. We should be raising WARNING, instead
of CRITICAL, based on the behaviour our previous plugin and the non-severity
of these checks. A zero value is passed to critical in order to disable the
default of 600s.