Remove `rbenv rehash` from `govuk_setenv`
This was failing (silently) because the global shims directory is owned by
root and `govuk_setenv` is typically called by the `deploy` user.

We shouldn't need this -- in theory. The only binaries that we need access
to are `ruby` and `bundle`. These are installed by `puppet-rbenv` as root.
I believe that everything after that is wrapped by `bundle exec`.