[#50887571] Elasticsearch local proxy
Description cribbed from puppetdoc:

Load balance connections to Elasticsearch by creating a loopback-only
vhost in Nginx which will forward to a set of Elasticsearch servers.

For the benefit of applications which use client libraries that don't
natively support load balancing. As a bonus, it will expose logging for
all requests to ES.

Where "applications" is "rummager" ;) Credit to @bobwalker and @samsharpe
for coming up with the solution.