Link to "low available disk" in ops manual
Currently it only contains a single section about `/boot` partitions. It
probably won't ever contain anything more - we can't guess all the other
types of partition-full scenarios.

But since we're seeing a lot of `/boot` related alerts at the moment, it
would be good to have the kernel cleanup docs to hand and not reinvent that
wheel.