[#50980739] Don't restart logstream unecessarily
Move logtream resources out of `govuk::app::service` to prevent logstream
processes from being restarted at the same time as apps. They shouldn't need
to be restarted when an app's config changes. Indeed, it's harmful to do so,
because we might miss important log entries during restarts.

I wasn't convinced that this warrants the complexity of a new sub-class.