[#50980739] Create mongodb::logging class
Move the creation of the log file and logrotate config to a single
sub-class. Define the variable for the log file path in a single place, and
pass this into both the `::logging` and `::config` classes, the latter of
which sets the path in the upstart job.

This will then enable us to move the logstream job out of `::service`.