[#50980739] Move logstream out of varnish::service
This will prevent the logstream service from being restarted when Varnish's
configuration changes, which isn't necessary.

Also includes some other monitoring related resources for the same reason.
The sub-class does not depend on anything, because we want monitoring and
logging to be setup regardless of the success of the class. However the
status of the sub-class is propogated out of `varnish::end`.