[#50737821] noop purge of unmanaged user accounts
Purge UNIX user accounts which are UID >= 500 and not managed by Puppet.
This will allow us to remove users from groups, or groups from environments,
and ensure that the systems are updated to reflect the manifests. Whereas
currently they are not.

At the moment a `noop => true` flag is placed on this resource so that it
can report the users that it matches without actually removing them. We
should review the output of this through Kibana to see if there are any
false-positives. Such as user accounts in the system UID range, or vice
versa. The noop can be removed once we're satisfied with the output.