Only set a default for UNICORN_HERDER_TIMEOUT when a 'rack' app
There is no idiom in Puppet for an optional parameter dependent on the
value of another parameter. To get around this, the common idiom in
the GOVUK Puppet repo is to use the 'NOTSET' string and have the
define's own internals hold the logic of how that value is set.

In this case, we only want to set a default UNICORN_HERDER_TIMEOUT
value when defining a 'rack' application. Hence the conditionals in
govuk::app::config and the default value logic provided in govuk::app.

There has been some discussion about whether it's worth having a
higher-level govuk::ruby_app define, but as it would be essentially
wrapping the the govuk::app define it was decided that this was more
explicit in terms of what's going on.