[#51241141] Fix unset var condition in spinup
Testing for an unset variable with `[[ $VAR ]]` is incompatible with
`set -u` further up this script, because bash will rightfully complain that
the variable is unbound.

Instead use shell parameter expansion[0] to substitute and match for given
string when `UNICORN_HERDER_TIMEOUT` is unbound. Using `:-` rather than `:=`
will not persist the substitution to the variable.

[0] http://www.gnu.org/software/bash/manual/bash.html#Shell-Parameter-Expansion