add server_name to nginx json log format
server_name is subtly different from http_host:

* http_host is the contents of the Host: header which the client
  specified
* server_name is the canonical name of the vhost which handled the
  request

there are a number of situations where they may differ. http_host may
contain uppercase characters; http_host may be empty or unrecognized
and route to a default vhost.

Therefore we should log server_name so we have a canonical
representation of the vhost which handled the request.