[#51148701] Pin all Elasticsearch versions
The following nodes have already been upgraded to 0.20 (from PPA) and no
longer need to use the extdata feature flag. We can pass a single version
string in direcly:

- `s_elasticsearch`
- `s_management_*`

The `s_development` Vagrant VM won't have been upgraded. Unless someone has
rebuilt their VM in the past couple of days (as explained below). But it
should be upgraded, to match the main ES cluster and CI.

The remaining nodes should still be running 0.19 (from our ghetto S3 repo)
until we choose to upgrade them in the near future. However I suspect that
the default param of `ensure => present` and the presence of the newer
package in the PPA, which is included on most nodes, will result in newly
provisioned machines being unexpectedly upgraded:

- `s_support`
- `s_logging_elasticsearch`

Passing a fixed version string in will maintain the status quo. We could
revert to using present/latest if we maintain a single package repo for each
environment in the future.