[#52126503] Fix redis-river install on ES 0.20
The installation of the redis-river plugin stopped working on Elasticsearch
0.20 due to some changes in the `plugin` command. This only affects new
installations - nodes upgrading from 0.19 to 0.20 will be unaffected. The
river appears to otherwise work fine on 0.20

In one rather innocuously titled commit, the mechanism for fetching plugins
hosted on GitHub was switched from the semi-deprecated "downloads" feature
where binaries could be uploaded to a project, to "zipball" URLs which
provide the contents of a repository at a given tag:

https://github.com/elasticsearch/elasticsearch/commit/2449049a8411f6ec0722c1cc054544a7eaf5938c

This means that the @leeadkins plugin can no longer provide a ZIP of the
built JARs through GitHub. They have to be uploaded to the Maven Central or
Sonatype repositories instead. I've asked him about this.

In the meantime though the old downloads link will still work and we can
pass a `-url` argument to force resolution. The `install_from` param has
been modified to accept a name or URL accordingly.

I've split the exec title/command out to ease testing. I'm not testing all
permutations of URL (HTTP, mixed case, etc) because Puppet is so slow. They
could be extended though.