Pull NetStorage 503 from internal static hostname
This is the template we serve from netstorage when applications
are dead. Static deliberately has a 10 year expiry. Templates
are not generally downloaded by the browser, they are served
through the application.

By grabbing these from Akamai, we are enforcing a 10 year expiry
on the Netstorage version of the page. When the template is updated,
we won't see the changes in Netstorage until 2023.

By switching to the internal non-akamai hostname for static, we
avoid this problem and the netstorage version of the 503 page will
be updated within a day of any updates to Static.