[#52201615] Added MockExtdata to add ability to update extdata
The earlier MockExtdata started using stubs from mocha after
introduction of puppetlabs_spec_helper. "stubs" doesn't supports blocks
because of which the hack to stub "File.exist?" started to fail.

To have the ability of update extdata for the existing test we are now
overriding extlookup function everytime we want to update extdata for
tests. This is a hack and should be removed when we move to hiera.