Add 'bouncer' app
bouncer is intended to replace the redirector. It is a simple rack app
that will serve redirects for the many hundreds of thounsands of urls
which we will need to redirect as part of the ALB transition.

Rather than putting bouncer on the redirector machines, we'll provision
new machine(s) for it configured as ruby app servers and then
decommission the redirector boxes once bouncer has taken over.