[#54183038] disable mysql slow log
A bunch of replication failures have been fixed recently by disabling
slow logs. This makes those changes stick.

Slow logs put extra locks in place, and can cause problems with
database replication. This commit disables slow logs in all
environments.

If we re-enable them, we should be sure we're actually going to use
the output of the slow log file, because it's relatively invasive to
the mysql process.