Revert "Merge pull request #806 from gds/use_statsd_metrics_in_nagios"
This reverts commit cc72a9be6f54d32e5d28b592312675b932b76881, reversing
changes made to 61bc93a9382bd8216c461ad04179bab45d10b0c9.

There is a problem with metrics which statsd doesn't know about yet. For
example, if you reboot a machine, then the http_5xx metric for a
particular app won't exist until that app has thrown an exception or
nginx has suffered a timeout, neither of which are particularly frequent
for smaller, less busy apps.

Reverting this commit while we think of a way to fix this problem.