Option for govuk::app::nginx_vhost to be default vhost
This introduces an option to govuk::app::nginx_vhost which is propogated
to nginx::config::vhost::proxy. If true, the listen directive will be
tagged with default_server, which makes that vhost the default for the
address:port pair.

I've taken care that no textual changes will occur to existing nginx
configs to avoid unneccesary app restarts when this manifest is
deployed. In particular avoiding an extra blank whitespace before the
semicolon of the listen directive.