[#54213626] Backup check shouldn't silent fail
Raise an UNKNOWN if unable to check the backup directories. Which I think
we'll find is all the time, because NRPE runs the check as UID `nagios` and
the directories are owned by `govuk-backup`. Currently an error is raised to
STDERR but an `OK` message and exit code 0 is used anyway. This will surface
the error for us to later fix - which may involve rewriting this plugin.