[#54213626] Passive checks for individual backups
Replace single existing Nagios check for backups with individual passive
checks for each backup. At the expensive of some extra code and check
resources, I think this is beneficial because:

- We can inspect the exit code of rsync directly. Rather than leaving a
  marker flag and then inferring.
- We can define exactly which backups to monitor rather than searching the
  filesystem for marker files.
- We don't need Nagios/NRPE to have file permissions on the backup targets,
  which was ultimately going to require root/sudo.
- We can let Nagios determine when it's been too long between backups.
  Currently set to 28hrs.

I've deliberately disabled the use of frequency periods other than `daily`
because I don't want to do the maths to work out the thresholds (how many
days are in a month?). We don't use these. If we come to use these in the
future then we should figure out the logic. I won't go as far as ripping all
the frequency support out though.

This still doesn't check the actual size of the resulting backup, which is a
more difficult problem to solve (how big do you expect the backup to be? in
total or incremental?). I think "did rsync run successfully?" is the more
pressing concern for now.