Omit Graphite on monitoring.management
This has since been split out to a separate `graphite-1.management` node. So
we no longer need to manage Graphite here.

It will continue to run. We can clean it up by hand if necessary. But newly
provisioned nodes, in Vagrant and vCloud, will no longer bring it up.

Adding an `include ::nginx` to Nagios since it was missing one of it's own.
It was silently relying on the Graphite module to provide it. Which resulted
in a broken dependency when Graphite was omitted.