[#54279372] Send logs to new logs-redis instances
Send JSON log events to the new pair of logs-redis instances. The rivers on
logs-elasticsearch are currently reading from both. So after a period of
time we should see a seamless switchover to the new pair. Then the old
rivers can be removed.

This will restart the logstream upstart jobs. After recent work to fix
dependencies it should NOT restart any other services.