[#54279372] Don't export rivers from old redis-N
Stop exporting river resources from the old `redis-N.backend` nodes now that
we are shifting traffic to `logs-redis-N.management`.

This will not remove the rivers. That should be done by hand when we're
happy that all logs events are going through the new redis nodes. However
this will stop them from being recreated in the future.