[#52225601] Manage puppet agents with Deb packages
To date the Puppet agents have been installed straight from gems during
provisioning and then not subsequently managed. Partly because we use a
package for the master and Puppet cannot manage a Deb and Gem of the same
name. This has resulted in:

- Upgrades being rolled "by hand" using Fabric.
- Modifying several places (vcloud-provisioner, Vagrant, etc) when we do
  change the agent version.
- Not really knowing what version is where. Not least because Rubygems will
  take the latest installed version of any package.

So it seems that the simplest solution, especially where bootstrapping new
or existing machines is concerned, is to use the Debian packages from
PuppetLabs. We do at least know that the apt/deb provider for `package{}` is
robust and can handle versions correctly. And it will match the existing
masters.

The versions of the master and agent will remain the same for now. If
required in future we could feature toggle the version for the
`puppetmaster` node. NB: This will also upgrade the master one minor
version, to match the current clients.

I'm removing the `/usr/local/bin` binstub that is created by the Rubygems
install. To prevent it taking precedence over the system package. Similarly
I'm also pushing out the `/usr/bin` binary to ensure that we have the right
one. When we come to uninstall the gems it will have to be done with
`--no-executables` to prevent these from being scrubbed.

This now includes the `apt.puppetlabs` repository everywhere. To reduce our
dependency on this we should spin a new story to mirror this, or specific
packages, onto our own infrastructure.