[#52225601] Make puppet agent work with rbenv
Modify the stock Puppet agent binstub to work with rbenv. The intepreter
uses the unqualified `ruby` to pick up rbenv's shim. `$LOAD_PATH` is
modified so that non-system Rubies can refer to the system-wide install of
Puppet in `vendor_ruby`.

This still requires some testing before we switch to rbenv in [#52292527].
Most notably of which:

- What happens to the version specific `ruby-json`?
- How do we handle `libshadow`?