Manage MySQL log rotation in puppet
Our mysql::server resources is parameterised with the location of the
MySQL error_log. The location we use (`/var/log/mysql/error.log`) is
not the standard Ubuntu location and is not managed by logrotate.

This leads to an issue where our error.log can grow in an unconstrained
fashion, consuming all the space on the server. It's also difficult to
manage as it can become quite large.

The only consumer of the error log is a govuk::logstream resource,
which runs with `tail -F` so will handle the log being rotated out from
under it with no issues.