check_reboot_required: don't use Exceptions as logic
For checking if files existed and were readable, this script was trying
to open the file, if it failed it would catch the exception and do
something.

Several people commented that this was less readable and that Exceptions
should not really be used for program flow. All of these have now been
converted to a standard `os.path.exists()` and if the path exists we
are assuming that it is readable as we are running as root.