[#44340317] Varnish check_http->check_tcp
Replace the Varnish HTTP check with a plain TCP check. This will allow it to
function in Preview, where BasicAuth is required.

I don't think that passing the BasicAuth credentials (conditional or
otherwise) is the right solution. Since the test is slightly
non-deterministic; it may serve a response from cache, or it might reverse
proxy an error directly from the backend. I did consider adding a synthetic
resource to the VCL and testing that - but that adds some complexity.

This will reduce the thoroughness of the check. We are no longer testing
whether Varnish is responding to HTTP requests, only that it's listening on
the TCP port. However all alerts that we've had from this service so far
have been `connect()` timeout errors. So it should make no functional
difference.