[#53865779] Monitoring wrapper for ext4mount
Wrap the `ext4mount` defined type with `govuk::mount` in order to DRY up
Nagios check exported resources.

An unfortunate lack of `**kwargs` means that we have to mirror parameters
one-for-one. This will need to be kept in step with the external module. We
can't use `undef` for `mountpoint` because it needs to default to `$title`.

And an unfortunate lack of exported resources in rspec-puppet means that we
can't actually test for Nagios resources.