[#47054095] Disable sshd agent forwarding
Agent forwarding is Bad because it allows other root users on the same
destination machine to access keys in your agent.

Previously we needed (and recommended) this to use fabric-scripts. This has
since been replaced by a remote method of using Fabric. We have amended all
`ssh_config` examples, but we can't be sure that someone-somewhere might be
using an old version.

By disabling it server-side we can ensure that it is no longer in use.
Except for the development VM, which people may bounce off of locally.