[#52292527] fix the sudo PATH problem properly
This fixes the same problem as fixed in 56fb4cb but better.

The fix in 56fb4cb is not really ideal as forwarding PATH is a
questionable behaviour for sudo. This commit instead hardcodes into
the sudoers file a secure_path variable which states what a sane
default path should be. We set the variable to include the rbenv path
so that we can run rbenv commands via sudo.