[#54832500] be explicit about who we are in our netstorage check
by using nagios_cache_bust rather than cache_bust we avoid colliding
with anyone else's cache busting. This means we can be more aware of
who is making this request in the logs, and also makes it slightly
less likely that we'll accidentally hit the cache from someone else's
request.