Make the mirror update script more informative
This commit changes the logic for two chained jobs from a simple `&&` to
some proper logic with informative log messages.

Currently our mirror is broken. Mirroring appears to be successful, but
there is no evidence that the upload to netstorage even starts. When
running the netstorage upload manually, it appears to be successful.

I have a suspicion that this chaining is failing:
    /usr/local/bin/govuk_update_mirror && \
        /usr/local/bin/govuk_update_mirror

That may be because the update_mirror script is returning with an
unsuccessful exit code, but we really can't tell.