[#44291865] Report stats on redis through collectd
We don't have the redis plugin included with collectd despite it being
available since v5.0. Instead, using a little Python script that uses
socket to make a connection.

Chosen this script instead of one that uses the redis python lib to
reduce the number of dependencies. At the moment we're just using this
to gather information on number of keys and objects in redis. If that
changes we _could_ move to using a script that uses the redis python
library.