[#51468317] Add a manifest to backup uploaded assets.
This installs a script and its dependencies to do a backup using duplicity to memset's memstore.
It will only do one full backup on first run (which has already happened). We wont need this since we intend to backup elsewhere eventually. We could of course add logic to do this if we wanted.