[#52506925] Create blockips.conf from puppet params
The nginx module now takes a list of IP addresses to block as the
parameter denied_ip_addresses. This is defaulted to an empty list.

The list of IP addresses is passed through to nginx::config where it
is used in conjunction with blockips.conf.erb to generate the block list
file.

A test for the generation of blockips.conf has been added to
nginx::config tests.