[#54836522] Add passive nagios check to mirrorer
Added a passive nagios check to init.pp in the mirrorer module called
check-mirrorer-${::hostname} with a description of mirrorer update and upload.

The govuk_update_and_upload script has been transformed into a template
that now includes a trap to catch EXIT and submit the exit code and a
message through to nagios.

The threshold has been set to 28hours.