Add carrierwave temp dir
Carrierwave is the library used by whitehall to manage file uploads. It
has a facility to temporarily store uploaded files before they are
attached to an activerecord model object and moved to their final
destination in the file system. By default these temporary files are
stored in RAILS_ROOT/public/uploads/tmp.

This has mostly been working for us, however that directory is not
propery managed and recently the folder was delted from preview causing
the temporary uploads to stop working.

Furthermore, in a load-balanced configuration, the temporary file needs
to be on a file system shared across all of the backend machines,
otherwise uploading from a temporary file will intermittently fail.

We will use this directory mounted on NFS to store the temporary
uploaded files.

I've not added a backup script for this path as the files really are
temporary files and do not need to be backed up.