[#55722254] Include libyajl1 dependency
When we switched from installing the collectd package to the
collectd-core package from apt the fact that libyajl1 was no longer
include was not caught. The reason for this is probably that no one has
built a fresh machine or vm using collectd and used curl_json so on the
machines in preview/staging/production the lib was still around from the
collectd pacakge.

libyajl1 is used by the curl_json module of collectd so the lack of it
was highlighted when working on #51962425