[#51962425] Add nagios check for syslog count
I've added a nagios graphite check that will monitor the derivative of
the syslog count and if this hits 0 trigger a critical. The reason that
the derivative is wrapped in a removeBelowValue(*, 0) is that when the
log indices are rolled the derivative will be plunged briefly to a
massively negative number and we don't want this to trigger a critical.

The check is done over a 60 second sliding window as this gives a better
response than the default of 5 minutes due. This is because the last
nagios check combined with requiring 5 minutes of absolutely 0 changes
in syslog messages can take upto 10 minutes, too long.