Don't use checksums to copy logs to the backup box.
This is useful when we're getting logs from Akamai, so we can make
doubly sure the logs came over correctly, but it's prohibitively
expensive to calculate checksums for all the archived logs on the log
machine and the backup machine. That, and we can be pretty confident
that we've transferred the files successfully between two of our
machines if the byte count matches.