Apparmor module has been merged upstream
I suddenly thought to PR our changes. I think I was putting it off because I
planned to write tests, but since we're using this module in govuk_mirror
now we might as well..

https://github.com/simondeziel/puppet-apparmor/pull/1

Update our fork to match. Will continue using the fork for resiliancy, in
case the upstream disappears for whatever reason. But switch back to master
branch retaining the same commit ref in the `.lock`.