Replace apt/unattended_upgrades with ext module
This replaces all of the functionality we currently have in two home grown
modules, with one that has more features and better support.

Retaining some quirks about our existing setup for compat. Such as running
`apt-get update` on every run and explicitly listing origins. We can review
these later. I'd also love to push the email address out to hiera, when we
have it.

The only difference that I can see is that we will no longer manage the
standard contents on `/etc/apt/sources.list`. But I don't feel this is
really necessary.

This means that we can now use other third-party modules which depend on the
apt module to install their own repos - like puppetlabs-mongodb.