[#54515428] Manage the MySQL debian-sys-maint user in Puppet
This user is created when MySQL is installed on Debian-based systems,
and its username and password are stored in /etc/mysql/debian.cnf.
Slaves need to have the same password as their master stored there,
and staging needs to have the same password as production. Preview
needs to have a different password from staging and production.

This commit sets a new password for this user on masters and slaves
from extdata, and manages /etc/mysql/debian.cnf, storing the new
password there. It only adds a new password in vagrant_credentials.csv
because I do not yet have access to credentials for the other
environments.